requirements.txt
# --- Data Ingestion & Cleaning ---
requests>=2.31.0
beautifulsoup4>=4.12.0
pypdf>=4.0.0

# --- Text Processing & Chunking ---
numpy>=1.24.0
nltk>=3.8.1        
tqdm>=4.66.0      

# --- Retrieval: Dense ---
torch>=2.2.0      
sentence-transformers>=2.5.0
faiss-cpu>=1.7.4   

# --- Retrieval: Sparse ---
bm25s>=0.1.7      

# --- Reader & Generation ---
transformers>=4.38.0
huggingface_hub>=0.20.0
openai>=1.14.0    

# --- Evaluation ---
scikit-learn>=1.4.0
rouge-score>=0.1.2 

selenium
webdriver-manager
ollama